+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [3.1.1] - 2023-09-17
+### Added
+* Log out a count of how many services were registered and a count by service lifetime.
+### Fixed
+* Corrupted/incompatible assemblies could cause the registration to not proceed due to an uncaught exception.
+
 ## [3.1.0] - 2023-04-16
 ### Fixed
 * Classes with no interface were not being correctly registered after some refactoring in 3.0.0